UDL: Fix uncheck buffer writing
> 679 main.c:263:33: error: '%s' directive output may be truncated writing up to 255 bytes into a region of size between 0 and 4095 [-Werror=format-truncation=]
> 680     snprintf(path, PATH_MAX, "%s/%s", dirname, dirent->d_name);
> 681                                  ^~
...
> 689 main.c:275:33: error: '%s' directive output may be truncated writing up to 255 bytes into a region of size between 0 and 4095 [-Werror=format-truncation=]
> 690     snprintf(path, PATH_MAX, "%s/%s", dirname, dirent->d_name);
> 691                                  ^~

+++ management/univention-directory-listener/src/main.c
@@ -260,7 +260,10 @@ static void purge_cache(const char *cache_dir) {
 	if ((dir = opendir(dirname)) != NULL) {
 		while ((dirent = readdir(dir))) {
 			char path[PATH_MAX];
-			snprintf(path, PATH_MAX, "%s/%s", dirname, dirent->d_name);
+			int ret;
+			ret = snprintf(path, PATH_MAX, "%s/%s", dirname, dirent->d_name);
+			if (ret < 0 || ret >= PATH_MAX)
+				abort();
 			unlink(path);
 		}
 		closedir(dir);
@@ -272,7 +275,10 @@ static void purge_cache(const char *cache_dir) {
 	if ((dir = opendir(dirname)) != NULL) {
 		while ((dirent = readdir(dir))) {
 			char path[PATH_MAX];
-			snprintf(path, PATH_MAX, "%s/%s", dirname, dirent->d_name);
+			int ret;
+			ret = snprintf(path, PATH_MAX, "%s/%s", dirname, dirent->d_name);
+			if (ret < 0 || ret >= PATH_MAX)
+				abort();
 			unlink(path);
 		}
 		closedir(dir);
